+++ src/main/java/org/bytedeco/javacv/OpenCVFrameGrabber.java
@@ -197,19 +197,22 @@ public class OpenCVFrameGrabber extends FrameGrabber {
         }
         cvSetCaptureProperty(capture, CV_CAP_PROP_CONVERT_RGB, imageMode == ImageMode.COLOR ? 1 : 0);
 
+        IplImage frame = null;
         try {
             // Before cvRetrieveFrame() starts returning something else then null
             // QTKit sometimes requires some "warm-up" time for some reason...
             // The first frame on Linux is sometimes null as well,
             // so it's probably a good idea to run this for all platforms... ?
             int count = 0;
-            while (count++ < 100 && cvGrabFrame(capture) != 0 && cvRetrieveFrame(capture) == null) {
+            while (count++ < 100 && cvGrabFrame(capture) != 0 && (frame = cvRetrieveFrame(capture)) == null) {
                 Thread.sleep(100);
             }
         } catch (InterruptedException ex) {
             // reset interrupt to be nice
             Thread.currentThread().interrupt();
-            return;
+        }
+        if (frame == null) {
+            throw new Exception("cvRetrieveFrame() Error: Could not retrieve frame in start().");
         }
 
         if (!triggerMode) {